+++ includes/dependencies.php
@@ -15,6 +15,7 @@ Autoloader::alias('Orchestra\Support\Form', 'Orchestra\Form');
 Autoloader::alias('Orchestra\Support\HTML', 'Orchestra\HTML');
 Autoloader::alias('Orchestra\Support\Memory', 'Orchestra\Memory');
 Autoloader::alias('Orchestra\Support\Messages', 'Orchestra\Messages');
+Autoloader::alias('Orchestra\Support\Site', 'Orchestra\Site');
 Autoloader::alias('Orchestra\Support\Table', 'Orchestra\Table');
 
 /*