+++ tests/macro.test.php
@@ -19,7 +19,8 @@ class MacroTest extends \Orchestra\Testable\TestCase {
 	 */
 	public function tearDown()
 	{
-		\URI::$uri = null;
+		\URI::$uri              = null;
+		\Orchestra\Site::$items = array();
 	}
 	
 	/**
@@ -33,11 +34,14 @@ class MacroTest extends \Orchestra\Testable\TestCase {
 		$memory = \Orchestra::memory();
 		$memory->put('site.name', 'Orchestra Test Suite');
 
-		$this->assertEquals('Orchestra Test Suite', \HTML::title(null));
-		$this->assertEquals('Home &mdash; Orchestra Test Suite', \HTML::title('Home'));
+		$this->assertEquals('Orchestra Test Suite', \HTML::title());
+
+		\Orchestra\Site::set('title', 'Home');
+
+		$this->assertEquals('Home &mdash; Orchestra Test Suite', \HTML::title());
 
 		$memory->put('site.format.title', ':page-title at :site-title');
-		$this->assertEquals('Home at Orchestra Test Suite', \HTML::title('Home'));
+		$this->assertEquals('Home at Orchestra Test Suite', \HTML::title());
 	}
 
 	/**
@@ -50,6 +54,34 @@ class MacroTest extends \Orchestra\Testable\TestCase {
 	{
 		$expected = '<?php foreach (Orchestra\Widget::make("placeholder."."foo")->get() as $_placeholder_): echo value($_placeholder_->value ?:""); endforeach; ?>';
 		$output   = \Blade::compile_string('@placeholder("foo")');
+
+		$this->assertEquals($expected, $output);
+	}
+
+	/**
+	 * Test blade compile @title
+	 * 
+	 * @test
+	 * @group core
+	 */
+	public function testBladeCompileTitle()
+	{
+		$expected = '<?php echo Orchestra\Site::get("title"); ?>';
+		$output   = \Blade::compile_string('@title');
+
+		$this->assertEquals($expected, $output);
+	}
+
+	/**
+	 * Test blade compile @description
+	 * 
+	 * @test
+	 * @group core
+	 */
+	public function testBladeCompileDescription()
+	{
+		$expected = '<?php echo Orchestra\Site::get("description"); ?>';
+		$output   = \Blade::compile_string('@description');
 
 		$this->assertEquals($expected, $output);
 	}