+++ views/account/index.blade.php
@@ -3,15 +3,9 @@
 @section('content')
 
 <div class="row-fluid">
-	<div class="span8">
-		<div class="page-header">
-			<h2>{{ ! empty($_title_) ? $_title_ : 'Something Awesome Without A Name' }}
-				@if ( ! empty($_description_))
-				<small>{{ $_description_ ?: '' }}</small>
-				@endif
-			</h2>
-		</div>
 
+	<div class="span8">
+		@include(locate('orchestra::layout.widgets.header'))
 		{{ $form }}
 	</div>
 
@@ -19,6 +13,7 @@
 		@placeholder('orchestra.account')
 		@placeholder('orchestra.helps')
 	</div>
+
 </div>
 
 @endsection