+++ views/extensions/configure.blade.php
@@ -4,13 +4,7 @@
 
 <div class="row-fluid">
 	<div class="span8">
-		<div class="page-header">
-			<h2>{{ ! empty($_title_) ? $_title_ : 'Something Awesome Without A Name' }}
-				@if ( ! empty($_description_))
-				<small>{{ $_description_ ?: '' }}</small>
-				@endif
-			</h2>
-		</div>
+		@include(locate('orchestra::layout.widgets.header'))
 
 		{{ $form }}
 	</div>