+++ views/resources/index.blade.php
@@ -3,17 +3,8 @@
 @section('content')
 
 <div class="row-fluid">
-	
-	<div class="page-header">
-		<h2>{{ ! empty($_title_) ? $_title_ : 'Something Awesome Without A Name' }} 
-			@if ( ! empty($_description_))
-			<small>{{ $_description_ ?: '' }}</small>
-			@endif
-		</h2>
-	</div>
-
+	@include(locate('orchestra::layout.widgets.header'))
 	{{ $table }}
-
 </div>
 
 @endsection